21.05.16 ========  + Improved performance of Resident Evil 8 when changing display settings,      it has a nasty habit of checking resolution support every frame and        SK now caches DXGI Factories and Resolution Mode Lists until they          actually change.
  ==> Bottom line, this saves up to 3 ms of DXGI API overhead every frame (!!)

 + Added texture mod support for NieR Sqrt 1.5
 + Added a drop-down menu in 'Display' to move borderless fullscreen windows
     between available monitors

+++ src/control_panel/cfg_compatibility.cpp
@@ -477,10 +477,14 @@ SK::ControlPanel::Compatibility::Draw (void)
 
         if (ImGui::IsItemHovered ())
           ImGui::SetTooltip ( _SummarizeWindowStyle (
-                                  SK_GetWindowLongPtrW (rb.windows.focus.hwnd, GWL_STYLE)
+                                  static_cast <DWORD> (
+                                    SK_GetWindowLongPtrW (rb.windows.focus.hwnd, GWL_STYLE)
+                                  )
                               ).c_str ());
 
-        ImGui::Text      ( "%8x", SK_GetWindowLongPtrW (rb.windows.focus.hwnd, GWL_EXSTYLE) );
+        ImGui::Text      ( "%8x", static_cast <DWORD> (
+                                    SK_GetWindowLongPtrW (rb.windows.focus.hwnd, GWL_EXSTYLE)
+                         )                            );
         ImGui::EndGroup ();
         if (rb.windows.focus.hwnd != rb.windows.device.hwnd)
         {
@@ -503,7 +507,9 @@ SK::ControlPanel::Compatibility::Draw (void)
           
           if (ImGui::IsItemHovered ())
             ImGui::SetTooltip ( _SummarizeWindowStyle (
-                                    SK_GetWindowLongPtrW (rb.windows.device.hwnd, GWL_STYLE)
+                                    static_cast <DWORD> (
+                                      SK_GetWindowLongPtrW (rb.windows.device.hwnd, GWL_STYLE)
+                                    )
                                 ).c_str ());
 
           ImGui::Text      ( "%8x", SK_GetWindowLongPtrW (rb.windows.device.hwnd, GWL_EXSTYLE) );
